requirements.txt
faiss-cpu==1.7.4
numpy==1.24.3
openai==1.0.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
tiktoken==0.4.0
langchain==0.0.208